refactor: CreatedAtをdomain/model/userからdomain/model/commonに移動したので修正した
